refactor: sync web funding URL with package.json and remove shared/funding
