--- shared/funding.ts
@@ -1,1 +1,0 @@
-export const FUNDING_URL = 'https://paypal.me/GehDoc';
